Move class to utils/__init__.py Put PRODUCT_NAME in a ProductName class similar to ProductType. Update usages of PRODUCT_NAME.
