+++ src/middlewared/middlewared/alert/source/auth.py
@@ -1,6 +1,6 @@
 from middlewared.alert.base import Alert, AlertCategory, AlertClass, AlertLevel, AlertSource
 from middlewared.alert.schedule import CrontabSchedule
-from middlewared.plugins.system.product import ProductType
+from middlewared.utils import ProductType
 from middlewared.utils.audit import UNAUTHENTICATED
 from time import time
