+++ src/middlewared/middlewared/alert/source/sensors.py
@@ -4,7 +4,7 @@
 # See the file LICENSE.IX for complete terms and conditions
 
 from middlewared.alert.base import AlertClass, AlertCategory, AlertLevel, AlertSource, Alert
-from middlewared.plugins.system.product import ProductType
+from middlewared.utils import ProductType
 
 
 class SensorAlertClass(AlertClass):
